Automatic text wrapping test
Implemented lu_StringWrap, a function that will scan over text and automatically word-wrap that text based on the display area and font being used. The function borrows logic from the game's own RenderText function.

Modified the Professor Birch intro to use lu_StringWrap for one of the longer pieces of text. In tests, it works!

I could apply this functionality to all typical dialogue and message displays to make it easier to author content -- having the game word-wrap your text so you don't have to do it manually (even though tools exist to automate much of that for you now). The main motivation, however, was for making improvements to battle text and flow -- coalescing multiple messages about single moves into just one or two messages, which the original devs couldn't do without automatic wrapping.

---

Documented some of my goals for this project in the `lu-docs` folder.

---

Set up a folder for quick-and-dirty experimental testing.