Game intro is faster; R&D/planning notes
Sped up the very beginning of the Professor Birch intro: the initial delay is 75% as long; Birch and the platform he stands on both fade in faster; and the delay between full opacity and the start of dialogue is shorter. Added documentation comments to some values and functions.

Created a "roadmap" document that I will probably not actually stick to, along with a companion document noting research into specific checklist items.

Updated the Custom Game planning document to elaborate on the new-game experience and the goals for the system's design.