Implemented the "toggle running" pref
Renamed fields in FieldInput to better represent their meanings: `pressedAButton` to `fieldButtonInteract`; and `pressedBButton` to `fieldButtonResurface`. These fields are only set when you're allowed to perform the actions in question.

Repurposed an unused bit in `FieldInput` to serve as `fieldButtonToggleRun`. Hooked it up to the "toggle running" feature, and modified top-level movement handlers to check and honor it.