Preliminary work on Custom Game options
Wrote C functions (based on OOP and templated C++ code I've written before) for serializing and parsing bitpacked data, and tested them in MSVC. They seem to work!

Wrote a struct definition for custom game options, along with serialization and parse functions. None of these options are hooked up to anything; nothing calls the serialization functions; so this data will just live in RAM with nothing ever touching it, for now.

Wrote a comment in the Littleroot Town scripts indicating where and how we should adjust the game intro flow based on whether our "start game with Running Shoes" option is enabled... once we actually implement it, that is.