Bitstream functions; cheat detection initial R&D
Added bitstream functions for u32s, s32s, strings (two variants), and buffers.

Fixed a mistake in the implementation of the s16 bitstream function.

Did some research into cheat detection. Implementing anti-cheat is not a goal of this ROM hack; however, as Action Replay and GameShark codes don't work on ROM hacks, it would be useful to be able to detect their usage and polirwlt warn the player of possible data corruption.