Savegame: PC is now bitpacked; fixes for inventory bitpacking
The Pokemon Storage System is now bitpacked.

Redid codegen for the inventory, to account for the fact that some bag pockets cap items at 999, not 99. Fixed how we handle item encryption in `item.c` to avoid corrupting item quantities.

With these changes, we save a total of 2555 bytes across the entire savegame, with the bulk of the savings in sectors used by SaveBlock1.

In-game tests establish that Box 1, at least, works perfectly when bitpacked.