Started work on Custom Game Options menu
Custom Game options are now reset when starting a new playthrough.

The way CGOs are stored in RAM has been updated. Unimplemented options are now clearly marked.

Wrote implementations for the following CGOs: start with running enabled; overworld poison rate/damage modifiers.

The old CGO menu implementation code has been reworked. Keybinds are now displayed in a strip along the top (same code as on the (nick)naming screen).

I think it'd be beneficial to split this into multiple H and C files. My rationale is that we already want a special UI for selecting Pokemon species as option values, so we're going to have layered screens/UI and we're going to need multiple sets of code files anyway.