CGO menu fixes; internal GUI documentation
Wrote some internal documentation how menus work.

Made fixes to the Custom Game Options menu. It now successfully keeps transparent pixels transparent, and is able to composite different parts of the menu... except for the controls/keybind strip, which still doesn't render. It's not even making it to VRAM, so even the text printing is going wrong. No idea why.