CGO menu works
Fixed a bug in my word-wrapping code, wherein if the last word in a string needed to wrap, it wouldn't, because we didn't process wrap points at the end of a line or the end of the string.

The Custom Game Options menu now has a working help feature. Additionally, the keybind strip now displays keybinds contextually depending on which menu item you have your cursor over (i.e. for submenus, or for things that have or do not have help text).

Testing confirms that the "start game with Running Shoes" option works. Poison options remain untested (rolled a bad Torchic and couldn't get far enough to test without grinding).