Attempting to add a scrollbar to the CGO menu
It doesn't really work, because there's something I failed to account for: sprites are scaled within a limited area. For example, an 8x8px sprite is scaled and rotated within an 8x8px area; if you use the double-size flag, it becomes a 16x16px area, which prevents rotation from causing cropping. In my case, I'm trying to resize a sprite (the scrollbar thumb) to well beyond twice its height, and that just won't work.

I might have to draw directly to the BG for this, if I can.