CGO menu scrollbar
Implemented a helper function for drawing a very simple scrollbar directly onto a BG layer. Unlike Game Freak's sprite-based scrollbars, we can display the scrollbar thumb at any size, to accurately indicate the overall size of the scrollable pane.

As scrollbars are window-relative, the options pane window in the CGO menu needed to be enlarged. It now uses the full screen width. This in turn meant that block offsets had to be changed for the help view, to avoid tiles getting clobbered.

The CGO menu now has its own palette.

The description for the "money loss on defeat" option was too long, and we don't have the ability to scroll help text in the CGO menu, so I've shortened it for now.