Item interaction script now coalesces multiple textboxes
Added new script commands for displaying message boxes with automatic word-wrapping.

Modified the built-in scripts used when you interact with items (visible or hidden) on the overworld, to coalesce multiple textboxes into one (i.e. "Player found one Item, and put it in the Category Pocket" or "Player found one Item, but the Bag is full"). Subjectively I feel this improves the pace of the game.