Fixed debug code
Turns out, including some of the C stdlib causes us to overshoot ROM space limits, and if we overshoot the 32MB limit, we compile a 128MB ROM.

Tests indicate that our bitpacking code is more-or-less working. Hard to be sure because the GBA debug-print code doesn't support advanced format specifiers (e.g. "%.7s") but the numeric values all look good.