Started work on integrating bitpacking into the savegame
Ported over my old edits to the savegame code, and updated them to match the new codegen.

Fixed bitpacking annotations and related changes on global structs and typedefs.

Disabled the unit test for bitpacking. If we run code generation multiple times within the ROM, even if in different translation units, we get naming conflicts (multiple symbols with the same name) and the build fails. This is because the generated code and variables are all extern (GCC's API sort of limits what we can do about that, so actually fixing it is a larger endeavor).

Saving is currently broken, however: sector ID 1 seems to be missing from the save file, so we can't load. Currently investigating.