lu-save-js: CViewElement: table cell tooltips
It's now possible for the canvas-rendered table cells to have tooltips, when their text-content is cut off. We generate DOM nodes to act as hitboxes. This is highly inefficient since we're spawning and destroying these nodes on every repaint, but the table still performs acceptably on desktop; I'll add a to-do list item to optimize this.