lu-save-js: Union fixes; value formatters
Fixed loading of union values: we no longer re-emplace the union value (clearing previous values), so the union is no longer empty; and we preserve the union tag and first members when we do the first emplace, which protects internally-tagged unions from losing their tag (between reading the tag and emplacing, should we emplace a different member than the one we used for reading the tag). Unfortunately, we do this for all unions, not just internally-tagged ones, because the bitpacking plug-in doesn't (yet) output the tag type (internal/external) for `c-type` elements. Additionally, we now properly select the right case node.

Fixed a mistake with the charset definitions.

CViewElement now displays strings as string literals, and now shows the string length in the Type column (e.g. `u8[7]` for a seven-char non-terminated string, or `u8[7 + sizeof('\0')` for a seven-char string that needs a terminator when in memory).

CViewElement: We now support defining type formatters; compare to writing DisplayString for a type in Natvis.

CViewElement: We now properly show the name of a union's active member.