lu-save-js fixes; bitpacking annotation fixes
lu-save-js: Fixed a mistake in the bitstream code that caused us to fail to consume characters when reading strings.

lu-save-js: Fixed a mistake in the SaveFormat code that caused us to misread the second save slot.

lu-save-js: CViewElement: When displaying strings, we no longer write in ligature names.

Emerald: Added or fixed bitpacking annotations on several values.