lu-save-js: Improvements to data representation; path tooltips
Created common base classes for "definition" versus "instance" classes.

"Instance" objects now know what they're a member of.

"Instance" objects now remember the definitions of the decl and type that are relevant to them; for example, a CStructInstance that represents the member of another struct now knows not only its type (CStruct), but also the decl (CValue) that defines it; and so it knows its own identifier (i.e. the member name).

CViewElement now unconditionally generates tooltips for item paths. The "Path" cells actually only show immediate identifiers, but the tooltips show the fully-qualified path.