lu-save-js: Started work on translations, charmap
Started work on a system for allowing user-defined translations between save formats, on a per-type basis. I've written up a new source file for it, but nothing uses that file yet.

To facilitate this, CStructInstance and CValueInstanceArray now offer member functions for rebuilding one of their members. This is mainly useful because the design of user-defined translators is such that they may clobber destination-aggregate members with a sentinel value. (Maybe we can preserve destination-aggregate members and simply restore them when that happens, instead of having to rebuild them?)

Started work on an alternate way to represent the game's character sets, this time with support for Japanese. The original goal was to facilitate comparisons between PokeStrings and JS string literals. Our current design for translation no longer needs this (at one point I wanted user-defined translators to be able to use bare JS objects, but this was impractical for a number of reasons) so I may or may not finish it or scrap it.