lu-save-js: Cleaned up the design of user-defined translators
Redesigned the behavior of user-defined translators. It's no longer valid to just clobber a CInstance (e.g. a member in a CStructInstance) with PASS; you must instead call this.pass on the CInstance, to recursively set the data inside of CInstances to PASS (i.e. it does myStructInstance.members[foo].value = PASS). The code for validating a successful user-defined translation checks that no CInstances were clobbered. This simplifies some other internal operations, simplifies the process by which we avoid clobbering the results of an outer translation when an inner translation PASSes on translating a nested value, and avoids the need to rebuild clobbered CInstances.

Default translation of integer values now validates that the source value can fit into the destination given the latter's bitpacking options.

CValue now offers a helper function for computing integer bounds based on bitpacking options. This is used by the default translation code and by CValueEditorElement.

Added a helper function to serve as an analogue to std::unreachable.