Update data-format-translator.js
When a default src-to-dst translation for a CValueInstance isn't possible, we instead write in the default value.

We no longer initialize CInstances to their defaults at the start of translation, as it'd be impossible to tell these defaults apart from values supplied by user-defined translations. This would prevent src-to-dst default translations. We do however still force an externally-tagged union to spawn the appropriate member, at the start of translation.