lu-save-js: Prep for checksum handling
Looked through the game's checksums and identified the various kinds: A32, A32U8, A16, and A32R16. (One data field is also checksummed using CRC16.) Expanded and renamed some of our checksum functions as prepwork; we eventually want to be able to auto-update checksums on structs as edits or translations are performed.