lu-save-js: Translation system fixes; re-save fixes
Fixed errors in the translation system.

It's now possible to register translators for top-level values, and to register translators for source versus destination CInstances.

Post-translation validity checks now enforce maximum lengths on buffers and strings, and do a better job of reporting error information for pointers. We also display primitive values as their JS representations (i.e. the string "foo" is displayed with quotes and escapes as needed).

Tested user-defined translators, with both valid translations and invalid ones. They work.

The InstructionsApplier is now more careful to write strings with the proper lengths, when saving. We pad strings as needed.

SaveSlot now properly sets `is_member_of` on top-level values. This makes it possible for the translation system to tell when it's dealing with a top-level value. (We could've just checked whether `is_member_of` is null/undefined, but this feels cleaner.)

Subclasses of AbstractDataFormatTranslator which override the wrong member name will now trigger errors when used.