lu-save-js: Union fixes
Fixed a typo in the union emplace functionality.

Default translation: if a destination union is internally tagged and empty, we check if its tag exists in the source union and if so, use that; this isn't new. What is now is that now, if the tag doesn't exist in the source union, we check if it has a default value within the destination union and if so, we use that.