lu-save-js: Critical fixes; offset/size info; category info
Upgraded the GCC bitpack plug-in to emit offset and size information for all struct members. This is something of a brute-force approach; we could make the output more compact by instead generating c-type elements for non-struct/non-union types e.g. u8 as well as any typedefs thereof.

Moved loading of basic type information from CStruct and CUnion to the base CTypeDefinition class. CValue now loads size/offset/bitfield information when it's present in the XML, and it also loads category names per value when they're present.

Fixed a critical issue in index.js that meant we were translating from the source format to the source format, rather than the selected destination format.

Implemented CBufferifier, a class that takes an aggregate CInstance and constructs a DataView reflecting how its contained data would be arranged in memory during the game's execution. Basic testing seems to work, but we can't easily verify the results. The goal for this is to aid in automatically recomputing checksums for those data structures in the savegame which use them, but our test save doesn't have any testcases and at present, we'd have to advance very far into the game to set any up.