lu-save-js: CViewElement now a subclass
CViewElement is now a TreeRowViewElement subclass with a built-in model. We take advantage of the new functionality for things like varying the text formatting within cells.

TreeRowViewElement now tells its model when it's asking for the text of a selected cell.

Fixed BBCode parsing mistakes involving the [raw][/raw] tag.