lu-save-js: TabViewElement
Created an element for tabbed views, in preparation for rebuilding the UI to work in terms of save files. (Each save file will be its own tab, with the save slots shown inside the tab body.) The goal is to make it so that instead of getting "one-shot" operations (e.g. "load this save file" or "load and then translate this save file"), you can load a save file, and then perform operations on it (including exporting it back to *.sav) at any time, including after having made manual changes to its data.