lu-save-js: Various critical fixes
Fixed handling of signed values.

Fixed how CViewElement displays CUnionInstances.

Fixed TreeRowViewElement to not choke when the browser can't be assed to actually update the computed styles we need to cache. This fixes a bug where the element would be stuck rendering empty content, because it'd be basing all of its positioning on NaNs.

Fixed loading of CUnionInstances: _emplace_for_load checks if the desired member is already active and if so, just returns it.