Checksum handling now in place; we can now save-edit BoxPokemon
CBufferifier is now able to remember and report the offset and size of a given value. Additionally, it now handles bitfields properly; it should be consistent with GCC.

Fixed an issue in CBufferifier where it wasn't handling the terminator byte on terminated strings.

CValueEditorElement now uses CBufferifier to auto-update checksums when editing a member of [a member of [...]] a struct that has one. This only works for checksums that have specific bitpacking categories (indicating the format of the checksum), that are not bitfields, and that are at the start or end of their immediate containing structs.

CValueEditorElement now allows the displayed CInstance path to word-wrap, and now uses a drop-down enum for booleans.

Edited SaveSlotElement to add special-case code to recalc BoxPokemon checksums. This code works: we can edit a Pokemon and have it show up properly in-game.

SaveSlots must now have an owning SaveFile, and are now constructed alongside that owning SaveFile.

TabViewElement now shows a fake tab when it has no tabs, as this looks less buggy than the alternative.

Fixed how Bitstream handles its offset.

---

Emerald: the SerializedBoxPokemonSubstructs checksum is no longer annotated as a checksum, as we need special-case logic for that one.