lu-save-js: Bitset editing UI and foundations for similar
Created BitsetInputElement, an element that offers a list of checkboxes for manipulating a bitset, while exposing the bitset as a single `value`.

Implemented CInstanceDisplayOverrides as a way of defining, on a SaveFormat, overrides for influencing the behavior of CViewElement and CValueEditorElement. Created a test override for displaying the flag bitsets in Emerald savedata (`emerald-display-overrides.js`).

Renamed CValuePath to InstructionNodeValuePath, and created a new CValuePath class for handling value paths on already-loaded data. The former class is used while loading and asserts correctness (because it's being applied to generated XML), while the latter class is more forgiving, more flexible, and reports errors in a friendlier manner. The new CValuePath additionally supports being applied to a CInstance (to resolve the path to a CInstance), and checking whether a CInstance matches the path (optionally relative to some base CInstance). The new CValuePath is used by CInstanceDisplayOverrides to check whether a given set of overrides should apply to a given CInstance.

Added a function for parsing BBCode into its plaintext content.