lu-save-js: Improvements to value display and editing
Updated CInstanceDisplayOverrides to support multiple (OR-linked) criteria for a single override collection.

Created a new custom element, EnumInputElement, to make it easier to use display overrides to display enums (i.e. if an enum lacks its own typedef, this change makes it easier to display the enum as an enum based on its path relative to its containing struct).

Added more Emerald-specific display overrides for certain enums.

When displaying arrays in the tree-row view, the array's own delimiters are no longer shown in green. Deemphsized text is also lighter.

Fixed a bug in CValueEditorElement where after a prior refactor, the means by which we looked up the applicable enum for a value was out of date.