lu-save-js: EnumInputElement improvements
EnumInputElement can now display as a drop-down or a textfield (the latter using a datalist), with a button to swap between the two. This should lead to a better editing experience on mobile (in theory anyway), where datalist just alters the suggestions shown in the keyboard, such that you can't necessarily see and scroll through all named values. It'll also lead to a better experience on Firefox for Desktop, which doesn't show datalist affordances at this time.