lu-save-js: TreeRowViewElement: Started work on column resizing
Started work on implementing resizable columns for TreeRowViewElement. We currently show a resize cursor in the right places, and allow changing the column widths. However, the resize code works by just forcing a column's width to an absolute value, rather than storing size changes as a delta, and this means that resizing the entire view causes things to fall apart a bit.