lu-save-js: Improvements to union editing
TreeRowViewElement now uses a WeakSet to track expanded items, at the very minor cost of no longer being able to treat primitives as items.

TreeRowViewElement now has APIs for setting the selected item, and for setting whether an item is expanded. These are used by the CValueEditorElement changes in this commit to ensure that when you expand a union, select its internal tag, and change that tag -- thereby replacing the contents of the union -- the union's active member is expanded as is consistent, and the replaced tag is selected.

CStructInstance now allows CValueInstances to know what union(s) they are an external tag for.

CValueEditorElement now detects when it's modifying a union tag, and updates the union(s) to emplace the appropriate member. That said, it doesn't prevent you from setting a value that doesn't correspond to a union member in each relevant union, and that's a problem.