lu-save-js-indexer: Started
Started work on an indexer for preprocessor enums and constants of interest. This would run as a post-build step whenever savedata is modified, to allow us to generate compact datafiles for these enums associated with each save format. The goal is to avoid having to hardcode things like Pokemon species names into lu-save-js.

The preprocessor is scripted in Lua, so Lua will likely end up being a requirement for this branch.