lu-save-js-indexer: Prototyping export of enum data
Wrote code which generates a binary blob representing the NATURE enum and the NUM_NATURES macro. In theory, we'd want to export this as "natures.dat". The current format for contiguous enums is to print length-prefixed strings, but I'm wondering if JS could parse delimited strings (maybe null-separated) faster. I should benchmark that before I proceed.

The dataview class now has a debugprint function.

The dataview class now properly validates offset-and-size pairs.

The dataview class now has functions to append a four-CC, an eight-CC, or a length-prefixed string (with an arbitrary prefix size).

Fixed various typos in the dataview code.