lu-save-js-indexer
Created an enumeration class to further aid in generating datafiles for enums.

Revised the datafile format for enums, in part based on JS benchmarking. Enumeration names are now serialized in a block, separated with a null byte. Multiple booleans in the output have also been combined into a single "flags" byte.