lu-save-js and lu-save-js-indexer
Fixed the indexer Lua script so that it no longer chokes when generating the datafile for an empty/irretrievable enum.

Our JavaScript code can now parse and display the content of the extra-data files, though a lot of the code for fetching the files and retrieving the parsed data is extremely flaky. (Bad error handling; bad/incomplete use of promises; etc..)

Dumped extra-data now includes contest information, script vars, and game stats. We also now dodge some macro names that aren't part of the enums we want, and various fixes have been made to generation of the ENUMDATA subrecord.

Fixed bugs in EnumInputElement re: updating the drop-down state when the value name set changes.

CInstanceDisplayOverrides can now override the displayed key for a CValueInstance inside of an array. That is: when array slots have meaningful names, you can use display overrides to display those names. We use this for script vars and game stats.