lu-save-js: Extra-data refactor
Refactored the JS code for loading and remembering extra-data. Each extra-data type now has its own class, so the "collection" and "index" classes don't need to do as much of the work for loading particular kinds of data, and can just focus on coordinating everything.

CInstanceDisplayOverrideCriteria is now constructible from a string, which is assumed to be a value path.

The post-build Lua script now properly outputs the CONTEST_CATEGORY enum.