lu-save-js and lu-save-js-indexer
lu-save-js: ExtraDataCollection now retains plain preprocessor constants. The current list consists of constants that may be of interest for a more advanced personality value editor in the future (should we find a way to emit the needed species info, which isn't available to the C preprocessor).

lu-save-js-indexer: Modified the core script to declare the files and data to output, and moved the functions that write subrecords to dataviews into their own file.