lu-save-js-indexer: We now copy the format XML to the right folder
We want lu-save-js to eventually "know" all of the savedata formats a project has historically used, as part of the functionality for converting a savegame between formats. This means that it's not enough to just have the current format's XML; the save editor needs to know past XMLs, so the post-build Lua script needs to copy a newly-generated XML file to a per-format folder.