lu-save-report-generator: Started work
Wrote a Lua post-build script (not yet hooked up to the build process) which parses the generated savedata format XML, and generates a human-readable Markdown-format report displaying statistics on the space usage and savings.

Fixed `stringify_table` so that the outermost table is designated #1, so that nested references to the outer table don't print it twice.