Fixes to save format indexer; save editor now uses the generated index
lu-save-js:

The save editor now reads the index.json file generated by our indexer. If you import a save file without supplying an XML dump of a savedata format, we'll pluck the version number out of your save file and choose the matching indexed format, if there is one.

Fixed reading of omitted values. Fixed printing of some extended characters by the DOMPokeStringPrinter. CArrayInstances of omitted DECLs now use "[omitted]" as their DisplayString.

The save editor UI now includes "clear" buttons for the file inputs, because for some unhinged reason browsers (*cough* Firefox) don't give the user any means whatsoever to clear a file input manually.

---

lu-save-js-indexer:

The Lua library now writes metatables to parsed JSON arrays, so you can tell whether a table was an array or an object within the source data. You can also use metafields to force a table to be serialized as an array, or to indicate that its Lua-side zero key is meaningful (and should be considered an array index).

When the save indexer loads index.json, if it sees that the format list is an array, it remaps the resulting Lua array so that format 0, if present, is preserved, and so that all other formats' (zero-based) version numbers are preserved.