Save editor: ENUNUSED subrecord
Created an ENUNUSED subrecord which can define ranges of unused values within an enum. The FLAG enum uses this to avoid serializing the names of the "FLAG_UNUSED_0x..." values, allowing us to shave 6KB off of `flags.dat`.

Fixed ExtraDataCollection so it won't choke if any of the enums it expects are missing. Fixed error logging for IndexedSaveFormatInfo.