Save editor: support for map information
The savedata format indexer now dumps the information needed for the save editor to understand map sections (i.e. Pokemon met-locations), map groups, and map numbers. Within the indexer, moved a lot of the code for retrieving and parsing macros into a helper function.

Fixed EnumInputElement to dispatch a bubbling InputEvent when the drop-down is changed; this allows CValueEditorElement to react to the drop-down changing.

Fixed CValueEditorElement so that if a display override refuses to create an element, we fall back to the default behavior.