lu_bitpack updates and similar
Save editor: Fixed how CValueEditorElement accesses the ChecksumRecalcHelper class.

Save editor: We now take advantage of the new game typedefs for map-related enums.

Save editor: Fixed application of CInstanceDisplayOverrideCriteria to integral typenames.

Updated the bitpack plug-in to fix issues with tracking integral types and to define a new attribute for annotations. This will allow us to keep things like "checksum" notes out of the bitpacking stats report. The save editor has been updated to load these as well.

Fixed the build script so that if a plug-in fails to build, we don't proceed with building the ROM or running the post-build Lua scripts.