Bitpacking updated; XML size reduced
Upgraded the bitpacking plug-in to reduce redundant information w.r.t. integral FIELD_DECLs, reducing the XML size by something like 16KB. This also fixes the sorting of typenames to be more consistent.

Updated the save editor to understand the new format changes. As part of this change, CIntegralTypeDefinitions are now created for typedefs of canonical integer types.

We now index and use the LANGUAGE enum in extra-data files and the save editor.