Improvement to the bitpack plug-in
Bitpack plug-in: Values that are declared as booleans (`bool8`) are now bitpacked as booleans, unless they are bitfields with a width greater than one bit (i.e. `SecretBase::toRegister`, which I assume Game Freak originally intended to use as a flags-mask or state enum).

The save editor has been updated to handle this change.