Work on game typedefs, map-num annotations
Documented the game typedefs.

Updated the PokemonLevel typedef to use the MAX_LEVEL constant.

Updated how the string length on `MauvilleOldManTrader::playerNames` is written, based on a look at the code. See the game typedefs documentation file for info.

Added an annotation to `SaveBlock1::outbreakLocationMapNum` to explicitly identify the associated map-group value. Updated the save editor to understand these annotations, such that we no longer need to hardcode the member names involved.