zero-fill-if-new plug-in-side support
Refactored the bitpack plug-in's XML generation, among other things, to allow bitpacking categories and annotations on structs and to output them optimally.

Added a new annotation, `zero-fill-if-new`, to the codebase. Support for it will be added to the save editor shortly.

Defaulted the Pokedex seen/caught masks to zero, to prep for dex extensions in the future.