Save editor fixes
Fixed a bug in EnumInputElement wherein if you spawned an element and then set its `value` to zero, nothing would happen and the element would appear empty. This was because the element only force-updates its state when that state changes, but we never initially set what the visible portions of the element display (i.e. failing to change it from its initial value caused this issue).

Fixed a mistake in the display overrides for map-num values.