ugh
Fixed a broken import in CViewElement.

Fixed an issue where Chromium et. al couldn't properly measure text in our virtual tree-row-table. MDN lists the properties we were using as "baseline" in a nice little green box... but with an asterisk that's a little too easy for the eyes to skip over. I kind of feel like this defeats the point of the whole "baseline" concept.